Add files via upload
The dataset shows the details of sales in a superstore. Columns include age, gender, profession, annual income, spending score etc. K means clustering is done to analyse the customer  behavioural patterns. Annual income and spending score are taken as the major variables which are highly contributing to the customer buying patterns.